requirements.txt
# LLM frameworks
langchain>=0.1.0
crewai>=0.28.0
autogen-agentchat>=0.2.0

# Open-source LLM support
openai>=1.0.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.20
ollama>=0.1.7

# Financial data APIs
yfinance>=0.2.18
alpha-vantage>=2.3.1
pandas-datareader>=0.10.0
polygon-api-client>=1.12.0

# Database handling and storage
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
alembic>=1.12.0

# ML libraries and data processing
tensorflow>=2.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
ta>=0.10.2  # Technical Analysis

# Visualization and dashboarding
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0
bokeh>=3.3.0

# Logging and monitoring
wandb>=0.16.0
mlflow>=2.8.0
loguru>=0.7.0

# Utils
python-dotenv>=1.0.0
schedule>=1.2.0
asyncio
aiohttp>=3.8.0
celery>=5.3.0
redis>=5.0.0